feat: :sparkles: useUserStore hook
the UserStore stores the current user and manages it through reducers such as updateUser, fetchUser, clearUser and setUser. This basically removes the direct user / role management from the AuthProvider . The AuthProvider  will be solely for the authentication. If in any screen we need to perform a certain function that will depend on the user, we will directly use the useUserStore  hook instead of having to first call the provider then get access to user details.